+++ app/src/main/res/layout/activity_music_temporary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".musicTemporary">
+    tools:context=".MusicTemporary">
 
     <TextView
         android:id="@+id/textWelcome"
@@ -45,10 +45,24 @@
         android:layout_marginStart="158dp"
         android:layout_marginTop="28dp"
         android:layout_marginEnd="159dp"
-        android:layout_marginBottom="352dp"
+        android:layout_marginBottom="29dp"
         android:text="Validate"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintBottom_toTopOf="@+id/textViewPlaying"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/musicName" />
+
+    <TextView
+        android:id="@+id/textViewPlaying"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="174dp"
+        android:layout_marginTop="15dp"
+        android:layout_marginEnd="179dp"
+        android:layout_marginBottom="267dp"
+        android:text="Playing: None"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/validate" />
 </androidx.constraintlayout.widget.ConstraintLayout>
